refactor(workout): replace modal with dedicated add workout page
This commit replaces the modal-based add workout functionality with a dedicated page for adding workouts. The changes include:
- Removing the `AddWorkoutContainer` and its related code.
- Introducing a new `AddWorkoutPage` for a more structured and maintainable approach to adding workouts.
- Updating navigation logic to use the new page instead of the modal.
- Refactoring related components to ensure consistency with the new design.

+++ lib/hive/adapter/calory_state.dart
@@ -5,7 +5,7 @@ part 'calory_state.g.dart';
 @HiveType(typeId: 2)
 class CaloryState extends HiveObject {
   @HiveField(0)
-  int calories;
+  double calories;
 
   @HiveField(1)
   DateTime timestamp;
@@ -13,7 +13,7 @@ class CaloryState extends HiveObject {
   CaloryState({required this.calories, required this.timestamp});
 
   CaloryState copyWith({
-    int? calories,
+    double? calories,
     DateTime? timestamp,
   }) {
     return CaloryState(
